Rename async and sync files
