+++ Src/Core/DatabaseCommand.Async.cs
